pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0", 
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "cleanome"
version = "0.1"
description = "Downloading genome annotations and turning them into genomics reference annotations"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.6"

# uncomment or adjust versions as you like:
dependencies = [
  "gtfparse",
  "polars-lts-cpu",
  "biopython",
  "pandas",
  "numpy",
  "tqdm",
  "requests",
  "ete3",
]

# metadata
[[project.authors]]
name = "Matthew Schmitz"

[project.scripts]
download_genomes         = "cleanome.download_genomes:main"
get_genomes_and_stats    = "cleanome.get_genomes_and_stats:main"
make_cellranger_arc_sh   = "cleanome.make_cellranger_arc_sh:main"
make_cellranger_rna_sh   = "cleanome.make_cellranger_rna_sh:main"
debug_gtf                = "cleanome.debug_gtf:main"
add_mito                 = "cleanome.add_mito:main"
